fix loi avatar
