+++ lib/src/ui/chat/conversation.dart
@@ -1,11 +1,17 @@
 import 'package:flutter/material.dart';
 
-
 class ConversationScreen extends StatelessWidget {
   const ConversationScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold();
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.black,
+        bottomOpacity: 0.0,
+        elevation: 0.0,
+      ),
+      body: Container(color: Colors.black),
+    );
   }
 }
